% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_strain_var.R
\name{add_strain_var}
\alias{add_strain_var}
\title{A function that will add variation among the strains within a functional group.
Variation is added to to the maximum growth rate and the inhibition strength.}
\usage{
add_strain_var(
  x,
  CB_var_gmax = 0,
  CB_var_h = 0,
  SB_var_gmax = 0,
  SB_var_h = 0,
  PB_var_gmax = 0,
  PB_var_h = 0,
  method = "even_v1.0"
)
}
\arguments{
\item{CB_var_gmax}{Amount of variability added to the g_max_CB parameter}

\item{CB_var_h}{Amount of variability added to the h_SR_CB parameter}

\item{SB_var_gmax}{Amount of variability added to the g_max_SB parameter}

\item{SB_var_h}{Amount of variability added to the h_O_SB parameter}

\item{PB_var_gmax}{Amount of variability added to the g_max_PB parameter}

\item{PB_var_h}{Amount of variability added to the h_O_PB parameter}

\item{method}{Method used to add strain variation.}

\item{An}{object returned by the `new_strain_parameter function`. Contains parameter values and starting abundance of strains and other state variables.}
}
\value{

}
\description{
A function that will add variation among the strains within a functional group.
Variation is added to to the maximum growth rate and the inhibition strength.
}
