% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bushplus_dynamic_model.R
\name{bushplus_dynamic_model}
\alias{bushplus_dynamic_model}
\title{The rate equations, as published in the paper, but with forcing of oxygen diffusivity \code{"a_0"} potential added,
and the possibility to simulate multiple strains per functional group}
\usage{
bushplus_dynamic_model(t, state, parameters)
}
\arguments{
\item{t}{The current time in the simulation}

\item{state}{A vector containing the current (named) values of each state variable}

\item{parameters}{A vector containing the current (named) values of each parameter}
}
\value{
Growth rate
}
\description{
The rate equations, as published in the paper, but with forcing of oxygen diffusivity \code{"a_0"} potential added,
and the possibility to simulate multiple strains per functional group
}
