% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bushplus_dynamic_model_strains_old.R
\name{bushplus_dynamic_model_strains_old}
\alias{bushplus_dynamic_model_strains_old}
\title{The rate equations, as published in the paper, but with forcing of oxygen diffusivity (a_0) potential added}
\usage{
bushplus_dynamic_model_strains_old(t, state, parameters)
}
\arguments{
\item{t}{The current time in the simulation}

\item{state}{A vector containing the current (named) values of each state variable}

\item{parameters}{A vector containing the current (named) values of each parameter}
}
\value{
Growth rate
}
\description{
The rate equations, as published in the paper, but with forcing of oxygen diffusivity (a_0) potential added
}
