% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_event_definition.R
\name{default_event_definition}
\alias{default_event_definition}
\title{A function that contains the events that should alter the state variables}
\usage{
default_event_definition(times = times, y = state, parms = parameters)
}
\arguments{
\item{times}{The time point in the simulation}

\item{parms}{Model parameters}

\item{state}{Current state variable values}
}
\value{
A vector of the state variables
}
\description{
A function that contains the events that should alter the state variables
}
