% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_definition_symmetric.R
\name{event_definition_symmetric}
\alias{event_definition_symmetric}
\title{Event definition for the symmetric simulation.}
\usage{
event_definition_symmetric(
  times,
  state,
  parms,
  log10aO_forcing_func,
  log10aS_forcing_func,
  noise_sigma,
  minimum_abundances
)
}
\arguments{
\item{times}{the time points in the simulation when events happen}

\item{state}{current state variable values}

\item{parms}{object of class \code{strain_parameter}}

\item{log10aO_forcing_func}{function to change oxygen diffusivity \code{aO}
depending on \code{t}}

\item{log10aS_forcing_func}{function. to change sulfide diffusivity \code{aS}
depending on \code{t}}

\item{noise_sigma}{value of variation added to \code{SO}, \code{SR}, \code{O}, and \code{P} using
the formula \code{rnorm(1, 0, noise_sigma * X) * noise_sigma * X}}

\item{minimum_abundances}{minimum abundances for CB, PB and SB. if the values
are lower, they will be set to these \code{minimum_abundances}}
}
\value{
the state vector, i.e. a vector containing the state variables.
}
\description{
This function contains events that can alter the state variables.
In this event definition, the abundances of all functional groups have 1 added to their density at each event.
}
