% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_expt_a_h_N.R
\name{get_final_states_a_h_N}
\alias{get_final_states_a_h_N}
\title{Function to get the steady states for combinations of a (oxygen diffusivity), inhibition strength and initial states.}
\usage{
get_final_states_a_h_N(x)
}
\arguments{
\item{expt}{A data frame of oxygen diffusivities and initial conditions combinations, for example created with expand.grid.}
}
\value{
Experimental result.
}
\description{
Function to get the steady states for combinations of a (oxygen diffusivity), inhibition strength and initial states.
}
