% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/get_stability_measures.replication_ssfind_result.R
\name{get_hysteresis_max}
\alias{get_hysteresis_max}
\title{Get the maximum of environmental conditions for which alternate stable states exist}
\usage{
get_hysteresis_max(up, down, a, threshold_diff)
}
\arguments{
\item{up}{State variable values as the environmental condition increases}

\item{down}{State variable values as the environmental condition decreases}

\item{a}{An environmental driver, here it is usually oxygen diffusivity}

\item{threshold_diff}{Amount of different between up and down states to qualify as alternate stable states}
}
\value{
A numeric value, which is the extent of the range of conditions for which alternate stable states exist.
}
\description{
Get the maximum of environmental conditions for which alternate stable states exist
}
