% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hysteresis_range.R
\name{get_hysteresis_range}
\alias{get_hysteresis_range}
\title{Get the range of environmental conditions for which alternate stable states exist}
\usage{
get_hysteresis_range(up, down, a)
}
\arguments{
\item{up}{State variable values as the environmental condition increases}

\item{down}{State variable values as the environmental condition decreases}

\item{a}{An environmental driver, here it is usually oxygen diffusivity}
}
\value{
Growth rate
}
\description{
Get the range of environmental conditions for which alternate stable states exist
}
