% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nonlinearity.R
\name{get_nonlinearity}
\alias{get_nonlinearity}
\title{Get the amount of non-linearity in a state-environment relationship.}
\usage{
get_nonlinearity(x, y)
}
\arguments{
\item{x}{Environmental driver state variable}

\item{y}{Ecosystem state variable}
}
\value{
Measure of non-linearity
}
\description{
Method described in Emancipator & Knoll (1993) Clinical Chemistry 39, 766-772,
\doi{10.1093/clinchem/39.5.766}
}
