% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stability_measures.R,
%   R/get_stability_measures.replication_ssfind_result.R,
%   R/get_stability_measures.temporal_ssfind_result.R
\name{get_stability_measures}
\alias{get_stability_measures}
\alias{get_stability_measures.replication_ssfind_result}
\alias{get_stability_measures.temporal_ssfind_result}
\alias{get_stability_measures.tbl_df}
\alias{get_stability_measures.data.frame}
\title{Get various measures of the stability}
\usage{
get_stability_measures(x, ...)

\method{get_stability_measures}{replication_ssfind_result}(ss_object)

\method{get_stability_measures}{temporal_ssfind_result}(ss_object)

\method{get_stability_measures}{tbl_df}(ss_object)

\method{get_stability_measures}{data.frame}(ss_object)
}
\arguments{
\item{x}{object from which to calculate the stability measures. At the moment \code{replication_ssfind_result} or \code{data.frame}.}

\item{...}{additional arguments for methods}
}
\value{
A data frame of stability measures of each state variable
}
\description{
The measures include non-linearity and hysteresis measures, of an ecosystem response to environmental change.
Takes steady state data as the input.
}
\examples{

}
