% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stability_measures_new.R
\name{get_stability_measures_new}
\alias{get_stability_measures_new}
\title{Version of function, working with data from the `run_temporal_ssfind_method` function. Gets various measures of the stability, specifically non-linearity and hysteresis measures, of an ecosystem response to environmental change.
Takes steady state data as the input.}
\usage{
get_stability_measures_new(ss_object)
}
\arguments{
\item{ss_object}{An object of class \code{ss_by_a_N_result} as returned by the ss_by_a_N() function or
the \code{result} e;ement of that object, i.e. \code{x$result}.}
}
\value{
A data frame of stability measures of each state variable
}
\description{
Version of function, working with data from the `run_temporal_ssfind_method` function. Gets various measures of the stability, specifically non-linearity and hysteresis measures, of an ecosystem response to environmental change.
Takes steady state data as the input.
}
