% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_symmetry_measurements.R
\name{get_symmetry_measurements}
\alias{get_symmetry_measurements}
\title{Computes measures that make comparisons between collapse and recovery
trajectory and between antagonistic environmental variables possible.}
\usage{
get_symmetry_measurements(res)
}
\arguments{
\item{res}{Results from \code{run_temporal_ssfind()} or from \code{run_temporal_ssfind_symmetric()}}
}
\value{
returns a frame containing symmetry measures for each variable, such as hysteresis area,
   shift magnitudes or distance between TP, as well as TP themselfes with their according state.
}
\description{
Computes measures that make comparisons between collapse and recovery
trajectory and between antagonistic environmental variables possible.
}
