% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth1.R
\name{growth1}
\alias{growth1}
\title{Growth rate function on one resource X}
\usage{
growth1(X, g_max, k_X)
}
\arguments{
\item{X}{Concentration of resource X}

\item{g_max}{Maximum growth rate}

\item{k_X}{Half saturation constant for resource X}
}
\value{
Growth rate
}
\description{
Growth rate function on one resource X
}
