% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bushplus_dynamic_model.R
\name{growth2}
\alias{growth2}
\title{Growth rate function on two resources X and Y}
\usage{
growth2(X, Y, g_max, k_X, k_Y)
}
\arguments{
\item{X}{Concentration of resource X}

\item{Y}{Concentration of resource Y}

\item{g_max}{Maximum growth rate}

\item{k_X}{Half saturation constant for resource X}

\item{k_Y}{Half saturation constant for resource Y}
}
\value{
Growth rate
}
\description{
Growth rate function on two resources X and Y
}
