% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inhibition.R
\name{inhibition}
\alias{inhibition}
\title{Growth inhibition function}
\usage{
inhibition(X, h_X)
}
\arguments{
\item{X}{Concentration of substance X}

\item{h_X}{Concentration of substance X at which the inhibition factor is 0.5 (i.e. the concerned rate is halved)}
}
\value{
Inhibition factor
}
\description{
Growth inhibition function
}
