% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microxanox-package.R, R/microxanox.R
\docType{package}
\name{microxanox-package}
\alias{microxanox}
\alias{microxanox-package}
\title{microxanox: Oxic-anoxic regime shifts in microbial communities}
\description{
The description will be discussed.

Ecosystems containing microbes can be oxic (oxyen present) or
anoxic (oxygen absent). These states can be both stable even with
identical biotic and abiotic circumstances, i.e. they can be
alternate stable states, implying history alone can determine the
present state. The biological reason is that the microbes can alter
the environmental conditions in a way that favour themselves, while
making them less suitable for other organisms. This creates mutual
inhibition. Which state occurs also depends on numerous
abiotic and biotic factors. This package facilitates simulation studies
of the influence of several of these factors.
}
\details{
Vignettes include:

- A user guide (maybe).
- A reproduction of some results of Bush et al 2017, on which the simulation is based. It shows the alternate stable states. It also extends the original study by adding temporal forcing and therefore switching between states, and also effects of biotic composition on the response to environmental change.
- A study of the effect of weaker or strong inbihibtion on the response to environmental change, include the extent of environmental conditions for which alternate stable states are present.
- A study of the effects of within functional group diversity on the response to environmental change.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/opetchey/microxanox/}
  \item Report bugs at \url{https://github.com/opetchey/microxanox/issues}
}

}
\author{
\strong{Maintainer}: Owen L. Petchey \email{Owen.Petchey@uzh.ch} (\href{https://orcid.org/0000-0002-7724-1633}{ORCID})

Other contributors:
\itemize{
  \item Rainer M. Krug \email{Rainer.Krug@uzh.ch} (\href{https://orcid.org/0000-0002-7490-0066}{ORCID}) [contributor]
}

}
\keyword{internal}
