% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialisation_functions.R
\name{new_initial_state}
\alias{new_initial_state}
\title{Create initial condition}
\usage{
new_initial_state(n_CB = 1, n_PB = 1, n_SB = 1, values = "bush_oxic")
}
\arguments{
\item{n_CB}{number of CB strains}

\item{n_PB}{number of PB strains}

\item{n_SB}{number of SB strains}

\item{values}{values to be used or \code{"bush"}, in which case the default from Bush et al (2017) will be used}
}
\value{
initial state \code{vector} as part of the strain_starter
}
\description{
Create initial condition
}
