% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_strain_parameter.R
\name{new_initial_state_anoxic}
\alias{new_initial_state_anoxic}
\title{Create initial condition}
\usage{
new_initial_state_anoxic(n_CB = 1, n_PB = 1, n_SB = 1, values = "bush")
}
\arguments{
\item{n_CB}{number of CB strains}

\item{values}{values to be used or \code{"bush"}, in which case the default from Bush et al (2017) will be used}

\item{N_PB}{number of PB strains}

\item{N_SB}{number of SB strains}
}
\value{

}
\description{
Create initial condition
}
