% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_replication_ssfind_parameter.R
\name{new_replication_ssfind_parameter}
\alias{new_replication_ssfind_parameter}
\title{Create parameter set to run a set of simulations to find stable states. Is passed to the function \code{run_replication_ssfind_parameter()} to run such a set of simulations.}
\usage{
new_replication_ssfind_parameter(...)
}
\arguments{
\item{...}{named parameter for the simulation to be set.
An error will be raised, if they are not part of the parameter set.}
}
\value{
object of the class \code{replication_ssfind_parameter}. This object of class
\code{replication_ssfind_parameter} is identical to an object of class \code{runsim_parameter}
plus an additional field \code{ss_expt} which contains a \code{data.frame} with
columns named
\itemize{
\item \code{N_CB}
\item \code{N_PB}
\item \code{N_SB}
\item \code{a_0}
which contain the initial states. If more than one strain is present, the same
initial state is assumed for each.
}
}
\description{
Create parameter set to run a set of simulations to find stable states. Is passed to the function \code{run_replication_ssfind_parameter()} to run such a set of simulations.
}
