% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_replication_ssfind_result.R
\name{new_replication_ssfind_results}
\alias{new_replication_ssfind_results}
\title{Create object of type \code{replication_ssfind_result} which is returned by the
function \code{run_replication_ssfind_parameter()}.}
\usage{
new_replication_ssfind_results(parameter, result)
}
\arguments{
\item{parameter}{object of class \code{replication_ssfind_parameter} which has been used
to run the simulation}

\item{result}{a \code{data.frame} containing the results of the simulation}
}
\value{
object of the class \code{replication_ssfind_result}. This object of class
\code{replication_ssfind_result} is identical to an object of class \code{replication_ssfind_parameter}
plus an additional field \code{result} which contains the result of the simulation.
}
\description{
Create object of type \code{replication_ssfind_result} which is returned by the
function \code{run_replication_ssfind_parameter()}.
}
