% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_runsim_result.R
\name{new_runsim_results}
\alias{new_runsim_results}
\title{Create result object of the function \code{run_sim()}.}
\usage{
new_runsim_results(parameter, result, help = FALSE)
}
\arguments{
\item{parameter}{object of class \code{runsim_parameter} which has been used to run the simulation}

\item{result}{a dataframe containing the results of the simulation}

\item{help}{if \code{TRUE}, the parameter will be listed and explained.}
}
\value{
parameter object of the class \code{runsim_result}
}
\description{
This object of class \code{runsim_result} is identical to the parameter object plus an additional field
  \code{result} which contains the result of the simulation.
}
