% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ssbyaN_parameter.R
\name{new_ss_by_a_N_parameter}
\alias{new_ss_by_a_N_parameter}
\title{Create parameter to run the simulation by using the function \code{ss_by_a_N()}}
\usage{
new_ss_by_a_N_parameter(..., help = FALSE)
}
\arguments{
\item{...}{named parameter for the simulation to be set. 
An error will be raised, if they are not part of the parameter set.}

\item{help}{if \code{TRUE}, the parameter will be listed and explained.}
}
\value{
parameter object of the class \code{ss_by_a_N_parameter}
}
\description{
Create parameter to run the simulation by using the function \code{ss_by_a_N()}
}
