% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_temporal_ssfind_result.R
\name{new_temporal_ssfind_results}
\alias{new_temporal_ssfind_results}
\title{Create object of type \code{temporal_ssfind_result} which is returned by the
function \code{run_temporal_ssfind_parameter()}.}
\usage{
new_temporal_ssfind_results(parameter = NULL, result)
}
\arguments{
\item{parameter}{not used - for consistency.}

\item{result}{a \code{data.frame} containing the results of the simulation}
}
\value{
the \code{result} object.
}
\description{
Create object of type \code{temporal_ssfind_result} which is returned by the
function \code{run_temporal_ssfind_parameter()}.
}
