% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_and_save.R
\name{plot_and_save}
\alias{plot_and_save}
\title{TODO}
\usage{
plot_and_save(
  CB_var_gmax,
  CB_var_h,
  SB_var_gmax,
  SB_var_h,
  PB_var_gmax,
  PB_var_h,
  sim_res,
  file_path_and_prefix
)
}
\arguments{
\item{CB_var_gmax}{TODO}

\item{CB_var_h}{TODO}

\item{SB_var_gmax}{TODO}

\item{SB_var_h}{TODO}

\item{PB_var_gmax}{TODO}

\item{PB_var_h}{TODO}

\item{sim_res}{TODO}

\item{file_path_and_prefix}{TODO}
}
\description{
TODO
}
