% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dynamics.R
\name{plot_dynamics}
\alias{plot_dynamics}
\title{Convenience function to plot the dynamics of a model run,
with strains within functional groups.}
\usage{
plot_dynamics(simulation_result, every_n = 1)
}
\arguments{
\item{simulation_result}{Object returned by the run_simulation function}

\item{every_n}{Plot data of every other n sample.}
}
\value{
returns the ggplot object of the plot. If it is assigned to a 
   variable, the plot needs to be plotted, otherwise it is plotted.
}
\description{
Convenience function to plot the dynamics of a model run,
with strains within functional groups.
}
