% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dynamics_strains.R
\name{plot_dynamics_strains}
\alias{plot_dynamics_strains}
\title{Convenience function to plot the dynamics of a model run,
with strains within functional groups}
\usage{
plot_dynamics_strains(simulation_result, every_n = 1)
}
\arguments{
\item{simulation_result}{Object returned by the run_simulation function.}

\item{every_n}{Plot data of every other n sample.}
}
\value{
Nothing.
}
\description{
Convenience function to plot the dynamics of a model run,
with strains within functional groups
}
