% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_symmetry_measures.R
\name{plot_symmetry_measures}
\alias{plot_symmetry_measures}
\title{Visualizes the symmetry measures obtained by the function \code{get_symmetry_measures()}}
\usage{
plot_symmetry_measures(res, species)
}
\arguments{
\item{res}{Results from \code{run_temporal_ssfind()} or from \code{run_temporal_ssfind_symmetric()}}

\item{species}{Environmental varibale of which the symmetry measures shoudl be plotted}
}
\value{
ggplot object, with shifts visualized as arrows and hysteresis area ribbon
}
\description{
Visualizes the symmetry measures obtained by the function \code{get_symmetry_measures()}
}
