% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectory_symmetry.R
\name{plot_trajectory_symmetry}
\alias{plot_trajectory_symmetry}
\title{Visualizes the symmetry of antagonistic collapse or recovery trajectories, symmetry measures of shift
are computed by the function \code{get_symmetry_measurements()}}
\usage{
plot_trajectory_symmetry(
  res,
  trajectory = "recovery",
  typ = "substrate",
  plot_log10 = FALSE
)
}
\arguments{
\item{res}{Results from \code{run_temporal_ssfind()} or from \code{run_temporal_ssfind_symmetric()}}

\item{trajectory}{Trajectory to be plotted. Either `collapse` or `recovery`.`.}

\item{typ}{Type of environmental variable. Either `substrate` or `bacteria`.}
}
\value{
ggplot object, with antagonistic shifts of related trajectory visualized as arrows.
}
\description{
Visualizes the symmetry of antagonistic collapse or recovery trajectories, symmetry measures of shift
are computed by the function \code{get_symmetry_measurements()}
}
