% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectory_symmetry_compact.R
\name{plot_trajectory_symmetry_compact}
\alias{plot_trajectory_symmetry_compact}
\title{Visualizes the symmetry of upards and downwards trajectories in the same figure, 
symmetry measures of shift are computed by the function \code{get_symmetry_measurements()}}
\usage{
plot_trajectory_symmetry_compact(res, typ = "substrate", plot_log10 = FALSE)
}
\arguments{
\item{res}{Results from \code{run_temporal_ssfind()} or from \code{run_temporal_ssfind_symmetric()}}

\item{typ}{Type of environmental variable. Either `substrate` or `bacteria`.}

\item{plot_log10}{Logical. If \code{TRUE}, logarithmic plot of the data is done.}
}
\value{
ggplot object, with antagonistic shifts of related trajectory visualized as arrows.
}
\description{
Visualizes the symmetry of upards and downwards trajectories in the same figure, 
symmetry measures of shift are computed by the function \code{get_symmetry_measurements()}
}
