% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_expt_a_N.R
\name{process_expt_result}
\alias{process_expt_result}
\title{A function to process the output of the ss_by_a_N (find steady states) function}
\usage{
process_expt_result(temp_result)
}
\arguments{
\item{temp_results}{The experimental results produced by ss_by_a_N}
}
\value{
Processed experimental results.
}
\description{
A function to process the output of the ss_by_a_N (find steady states) function
}
