% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_expt_result.R
\name{process_ss_result}
\alias{process_ss_result}
\title{A function to process the output of the \code{"get_final_states"} function}
\usage{
process_ss_result(ss_expt, temp_result)
}
\arguments{
\item{ss_expt}{TODO}

\item{temp_result}{The experimental results produced by ss_by_a_N}
}
\value{
Processed experimental results.
}
\description{
A function to process the output of the \code{"get_final_states"} function
}
