% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_anoxic_1strain_initial_state.R
\name{return_anoxic_1strain_initial_state}
\alias{return_anoxic_1strain_initial_state}
\title{Function that returns default initial states.
This particular set favours the anoxic state.}
\usage{
return_anoxic_1strain_initial_state()
}
\value{
A named vector of parameter values
}
\description{
Function that returns default initial states.
This particular set favours the anoxic state.
}
