% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_bushetal2017_parameter_values.r
\name{return_bushetal2017_parameter_values}
\alias{return_bushetal2017_parameter_values}
\title{Function that returns default parameter value set.
This is the set used in Bush et al (2017) to give a system
that can display alternate stable states.}
\usage{
return_bushetal2017_parameter_values()
}
\value{
A named vector of parameter values
}
\description{
Function that returns default parameter value set.
This is the set used in Bush et al (2017) to give a system
that can display alternate stable states.
}
