% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_replication_ssfind.R
\name{run_replication_ssfind}
\alias{run_replication_ssfind}
\title{Run simulations to determine the steady states by the replication method.}
\usage{
run_replication_ssfind(parameter, mc.cores = getOption("mc.cores", 0))
}
\arguments{
\item{parameter}{object of class \code{replication_ssfind_parameter} as generated by \code{new_replication_ssfind_parameter()}}

\item{mc.cores}{the number of cores to be used. If 0, the old sequential version is used.
The default is read from the option \code{mc.cores}, i.e. using \code{getOption("mc.cores", 0)}.}
}
\value{
Processed data about steady states
}
\description{
Function to get the steady states for combinations of \code{a} (oxygen diffusivity) and initial states.
This function is multithreaded, and the value of mc.cores determines the umber of parallel threads.
}
