% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Run the simulation}
\usage{
run_simulation(parameter)
}
\arguments{
\item{parameter}{an object of class \code{runsim_parameter} as returned by
\code{new_runsim_parameter()}.}
}
\value{
an object of class 'runsim_result\verb{, obtained from running the su=imulation as defined in }parameter\code{.}
}
\description{
Run the simulation
}
