% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation_symmetric.R
\name{run_simulation_symmetric}
\alias{run_simulation_symmetric}
\title{Run the symmetric simulation}
\usage{
run_simulation_symmetric(parameter)
}
\arguments{
\item{parameter}{an object of class \code{runsim_parameter} as returned by
\code{new_runsim_parameter()}. Needs to hold key sym_axis.}
}
\value{
an object of class \code{runsim_result}, obtained from running the
simulation as defined in \code{parameter}.`
}
\description{
This function takes the \code{parameter} object and runs a simulation based on these.
It returns an object of class \code{runsim_result} which contains an additional
entry, i.e. \code{result} which contains the results of the simulation. The
simulation can be re-run using the returned object as input \code{parameter}.
}
