% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_temporal_ssfind.R
\name{run_temporal_ssfind}
\alias{run_temporal_ssfind}
\title{Finds the stable states for a parameter set by increasing and decreasing
the oxygen diffusivity in a stepwise fashion.}
\usage{
run_temporal_ssfind(parameter)
}
\arguments{
\item{parameter}{an object of class \code{runsim_parameter} as returned by
\code{new_runsim_parameter()}.}
}
\value{
A data frame of final states and oxygen diffusivity values
}
\description{
Finds the stable states for a parameter set by increasing and decreasing
the oxygen diffusivity in a stepwise fashion.
}
