% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_temporal_ssfind_method.R
\name{run_temporal_ssfind_method}
\alias{run_temporal_ssfind_method}
\title{Find the stable states for a parameter set}
\usage{
run_temporal_ssfind_method(p)
}
\arguments{
\item{parameter}{an object of class \code{runsim_parameter} as returned by
\code{new_runsim_parameter()}.}
}
\value{
an object ...
}
\description{
Find the stable states for a parameter set
}
