% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_temporal_ssfind_symmetric.R
\name{run_temporal_ssfind_symmetric}
\alias{run_temporal_ssfind_symmetric}
\title{Used to find the stable states for a parameter set by increasing and decreasing
the oxygen diffusivity in a stepwise fashion. If increasing \code{parameter$sim_duration}
while keeping the length of \code{parameter$log10_series} doesn't change response dynanimcs,
stable states have been found.}
\usage{
run_temporal_ssfind_symmetric(parameter)
}
\arguments{
\item{parameter}{an object of class \code{runsim_parameter} as returned by
\code{new_runsim_parameter()}.}
}
\value{
A data frame of final states, oxygen and sulfide diffusivity values
}
\description{
Used to find the stable states for a parameter set by increasing and decreasing
the oxygen diffusivity in a stepwise fashion. If increasing \code{parameter$sim_duration}
while keeping the length of \code{parameter$log10_series} doesn't change response dynanimcs,
stable states have been found.
}
