% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_diffusivities.R
\name{set_diffusivities}
\alias{set_diffusivities}
\title{Used to update the asymmetry factor with logaO_series & log10aS_series accordingly}
\usage{
set_diffusivities(param, asym_factor = 1)
}
\arguments{
\item{param}{:parameter set of type \code{runsim_parameter}}

\item{asym_factor}{:asym_factor to update \code{runsim_parameter}}
}
\value{
updated parameterset
}
\description{
The \code{set_diffusivities} enables manipulating \code{aS} forcing in asymmetric manner to decrease (<1) or increase (>1) stress on cyanobacteria. If the vector in
\code{parameter$log10a_series} has multiple local minima, maxima respectively, it  is directly mirrored. ATTENTION: Mean value of the series has to be the axis at which the
vector is mirrored!
}
