% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_expt_a_h_N.R
\name{ss_by_a_h_N}
\alias{ss_by_a_h_N}
\title{Function to get the steady states for combinations of a (oxygen diffusivity), inhibition strength, and states.}
\usage{
ss_by_a_h_N(expt)
}
\arguments{
\item{expt}{A data frame of oxygen diffusivities and initial conditions combinations, for example created with expand.grid.}
}
\value{
Experimental result.
}
\description{
Function to get the steady states for combinations of a (oxygen diffusivity), inhibition strength, and states.
}
