% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_temporal_symmetric.R
\name{sym_plot_temporal_ss}
\alias{sym_plot_temporal_ss}
\title{Plot the temporal dynamics of a model run of a symmetric parameter set}
\usage{
sym_plot_temporal_ss(temporal_results, zoom = FALSE)
}
\arguments{
\item{temporal_results}{Results from \code{run_temporal_ssfind()}}

\item{zoom}{If \code{TRUE}, zoommin to an area (only as an example - no real use)}
}
\value{
returns the ggplot object of the plot. If it is assigned to a 
   variable, the plot needs to be plotted, otherwise it is plotted.
}
\description{
This is a convenience function to plot the dynamics of a model run particularly
for a symmetric model
with strains within functional groups displayed.
}
