% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_strain_var.R
\name{add_strain_var}
\alias{add_strain_var}
\title{Create Variability in strain parametern}
\usage{
add_strain_var(
  x,
  CB_var_gmax = 0,
  CB_var_h = 0,
  SB_var_gmax = 0,
  SB_var_h = 0,
  PB_var_gmax = 0,
  PB_var_h = 0,
  method = "even_v1.0"
)
}
\arguments{
\item{x}{}

\item{CB_var_gmax}{}

\item{CB_var_h}{}

\item{SB_var_gmax}{}

\item{SB_var_h}{}

\item{PB_var_gmax}{}

\item{PB_var_h}{}

\item{method}{}
}
\value{

}
\description{
TO BE DOCUMENTED!!!!!
}
