% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bushplus_dynamic_model.R
\name{bushplus_dynamic_model}
\alias{bushplus_dynamic_model}
\title{The rate equations, as published in the Bush et al 2017
\doi{10.1093/clinchem/39.5.766} paper, but with forcing of
oxygen diffusivity \code{a_0} potential added, and the possibility to simulate
multiple strains per functional group}
\usage{
bushplus_dynamic_model(t, state, parameters, log10a_forcing_func, ...)
}
\arguments{
\item{t}{The current time in the simulation}

\item{state}{A vector containing the current (named) values of each state
variable}

\item{parameters}{An object of class \code{runsim_parameter} as returned by
`new_runsim_parameter()``}

\item{log10a_forcing_func}{function to change oxygen diffusivity \code{a} depending on \code{t}}

\item{...}{not used. Needed to catch additional parameters.}
}
\value{
a list containing two elements, namely the rate of change of the
strains, and also the current values of oxygen diffusivity \code{a}.
}
\description{
The rate equations, as published in the Bush et al 2017
\doi{10.1093/clinchem/39.5.766} paper, but with forcing of
oxygen diffusivity \code{a_0} potential added, and the possibility to simulate
multiple strains per functional group
}
