% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bushplus_dynamic_model.R
\name{bushplus_dynamic_model}
\alias{bushplus_dynamic_model}
\title{The rate equations, as published in the Bush et al 2017 paper, but with forcing of oxygen diffusivity \code{"a_0"} potential added,
and the possibility to simulate multiple strains per functional group}
\usage{
bushplus_dynamic_model(t, state, parameters)
}
\arguments{
\item{t}{The current time in the simulation}

\item{state}{A vector containing the current (named) values of each state variable}

\item{parameters}{A list containing the parameter values}
}
\value{
A list containing the rate of change of each state variable, and also the current values of oxygen diffusivity.
}
\description{
The rate equations, as published in the Bush et al 2017 paper, but with forcing of oxygen diffusivity \code{"a_0"} potential added,
and the possibility to simulate multiple strains per functional group
}
