% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_parameter_set.R
\name{construct_parameter_set}
\alias{construct_parameter_set}
\title{Construct set of parameter values}
\usage{
construct_parameter_set(
  CB_strain_params,
  PB_strain_params,
  SB_strain_params,
  all_other_parms
)
}
\arguments{
\item{all_other_parms}{Vector of all other parameters}

\item{CB_strain_parms}{Data frame of parameter values of CB strains}

\item{PB_strain_parms}{Data frame of parameter values of PB strains}

\item{SB_strain_parms}{Data frame of parameter values of SB strains}
}
\value{
A vector of a full set of parameter values
}
\description{
Construct set of parameter values
}
