% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_definition_1.R
\name{event_definition_1}
\alias{event_definition_1}
\title{A function that contains the events that should alter the state variables}
\usage{
event_definition_1(times = times, y = state, parms = parameters)
}
\arguments{
\item{times}{The time point in the simulation}

\item{y}{Current state variable values}

\item{parms}{Model parameters}
}
\value{
A vector of the state variables
}
\description{
A function that contains the events that should alter the state variables
}
