% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_definition_1.R
\name{event_definition_1}
\alias{event_definition_1}
\title{A function that contains events that should alter the state variables.
(It is named definition "1", in case other definitions are created.)}
\usage{
event_definition_1(
  times = times,
  y = state,
  parms = parameters,
  log10a_forcing_func,
  noise_sigma,
  minimum_abundances
)
}
\arguments{
\item{times}{the time point in the simulation}

\item{y}{current state variable values}

\item{parms}{object of class \code{strain_parameter}}

\item{log10a_forcing_func}{TODO}

\item{noise_sigma}{TODO}

\item{minimum_abundances}{TODO}
}
\value{
a vector of the new state variables
}
\description{
A function that contains events that should alter the state variables.
(It is named definition "1", in case other definitions are created.)
}
