% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_final_states_a_N.R
\name{get_final_states_a_N}
\alias{get_final_states_a_N}
\title{Get the steady state solutions for a set of oxygen diffusivity and initial states.}
\usage{
get_final_states_a_N(x, parameter)
}
\arguments{
\item{x}{A vector of oxygen diffusivity and initial conditions.}

\item{parameter}{object of class \code{ss_by_a_N_parameter} as generated by \code{new_ss_by_a_N_parameter()}}
}
\value{
A vector of steady states
}
\description{
Can be used directly. Also can be used indirectly, via the function \code{ss_by_a_N()}
}
