% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stability_measures.R
\name{get_hysteresis_max}
\alias{get_hysteresis_max}
\title{Get the range of environmental conditions for which alternate stable states exist}
\usage{
get_hysteresis_max(up, down, a)
}
\arguments{
\item{up}{State variable values as the environmental condition increases}

\item{down}{State variable values as the environmental condition decreases}

\item{a}{An environmental driver, here it is usually oxygen diffusivity}
}
\value{
A numeric value, which is the extent of the range of conditions for which alternate stable states exist.
}
\description{
Get the range of environmental conditions for which alternate stable states exist
}
