% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stability_measures.R
\name{get_hysteresis_total}
\alias{get_hysteresis_total}
\title{Get the total hysteresis of a system variable.}
\usage{
get_hysteresis_total(up, down)
}
\arguments{
\item{up}{State variable values as the environmental condition increases}

\item{down}{State variable values as the environmental condition decreases}
}
\value{
A numeric value--the total hysteresis.
}
\description{
Get the total hysteresis of a system variable.
}
