% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stability_measures.R
\name{get_stability_measures}
\alias{get_stability_measures}
\title{Get various measures of the stability of an ecosystem response to environmental change.
Takes steady state data as the input.}
\usage{
get_stability_measures(ss_object)
}
\arguments{
\item{ss_object}{An object of the type returned by the ss_by_a_N() function}
}
\value{
A data frame of stability measures of each state variable
}
\description{
Get various measures of the stability of an ecosystem response to environmental change.
Takes steady state data as the input.
}
