% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stability_measures.R,
%   R/get_stability_measures.replication_ssfind_result.R,
%   R/get_stability_measures.temporal_ssfind_result.R
\name{get_stability_measures}
\alias{get_stability_measures}
\alias{get_stability_measures.replication_ssfind_result}
\alias{get_stability_measures_replication_ssfind_result}
\alias{get_stability_measures.temporal_ssfind_result}
\alias{get_stability_measures_temporal_ssfind_result}
\title{Get various measures of the stability}
\usage{
get_stability_measures(ss_object, threshold_diff_log10scale, ...)

\method{get_stability_measures}{replication_ssfind_result}(ss_object, threshold_diff_log10scale = 3, ...)

get_stability_measures_replication_ssfind_result(ss_object)

\method{get_stability_measures}{temporal_ssfind_result}(ss_object, threshold_diff_log10scale = 3, ...)

get_stability_measures_temporal_ssfind_result(ss_object)
}
\arguments{
\item{ss_object}{object from which to calculate the stability measures. At
the moment \code{replication_ssfind_result} or \code{data.frame}.}

\item{threshold_diff_log10scale}{Amount of different between up and down
states to qualify as alternate stable states, on log10 scale}

\item{...}{additional arguments for methods}
}
\value{
A data frame of stability measures of each state variable
}
\description{
The measures include non-linearity and hysteresis measures, of an ecosystem
response to environmental change. Takes steady state data as the input.
}
