% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_strain_parameter.R
\name{new_CB_strain_parameter}
\alias{new_CB_strain_parameter}
\title{Create empty CB strain parameter}
\usage{
new_CB_strain_parameter()
}
\value{
data.frame with additional class \code{CB_strain_parameter}
}
\description{
Create empty CB strain parameter
}
