% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_initial_state.R
\name{new_initial_state}
\alias{new_initial_state}
\title{Create initial state of the system}
\usage{
new_initial_state(n_CB = 1, n_PB = 1, n_SB = 1, values = "bush_anoxic_fig2ab")
}
\arguments{
\item{n_CB}{number of CB strains}

\item{n_PB}{number of PB strains}

\item{n_SB}{number of SB strains}

\item{values}{Allowed values are:
\itemize{
\item \code{"bush_anoxic_fig2ab"}: initial conditions for Bush et al 2017 figure 2 a & b
\item \code{"bush_oxic_fig2cd"}: initial conditions for Bush et al 2017 figure 2 c & d
\item \code{"bush_ssfig3"}: initial conditions for Bush et al 2017 figure e
\item \code{"NA"}: all initial values set to \code{NA}. Usable for e.g. setting own initial state
}}
}
\value{
initial state \code{vector} aof the system s part of the strain_starter
}
\description{
Create initial state of the system
}
