% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_strain_parameter.R
\name{new_PB_strain_parameter}
\alias{new_PB_strain_parameter}
\title{Create PB strain parameter}
\usage{
new_PB_strain_parameter(n = 1, values = "bush")
}
\arguments{
\item{n}{number of strains}

\item{values}{values to be used or \code{"bush"}, in which case the default from Bush et al (2017) will be used}
}
\value{
data.frame with additional class \code{PB_strain_parameter}
}
\description{
Create PB strain parameter
}
