% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_runsim_result.R
\name{new_runsim_results}
\alias{new_runsim_results}
\title{Create object of type \code{runsim_result} which is returned by the
function \code{run_sim()}.}
\usage{
new_runsim_results(parameter, result)
}
\arguments{
\item{parameter}{object of class \code{runsim_parameter} which has been used
to run the simulation}

\item{result}{a \code{data.frame} containing the results of the simulation}
}
\value{
object of the class \code{runsim_result}. This object of class
\code{runsim_result} is identical to an object of class \code{runsim_parameter}
plus an additional field \code{result} which contains the result of the simulation.
}
\description{
Create object of type \code{runsim_result} which is returned by the
function \code{run_sim()}.
}
