% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_strain_parameter.R
\name{new_SB_strain_parameter}
\alias{new_SB_strain_parameter}
\title{Create empty SB strain parameter}
\usage{
new_SB_strain_parameter()
}
\value{
data.frame with additional class \code{SB_strain_parameter}
}
\description{
Create empty SB strain parameter
}
