% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_SB_strain_parameter.R
\name{new_SB_strain_parameter}
\alias{new_SB_strain_parameter}
\title{Create new \code{SB_strain_parameter} object}
\usage{
new_SB_strain_parameter(n = 1, values = "bush")
}
\arguments{
\item{n}{number of strains}

\item{values}{Allowed values are:
\itemize{
\item \code{"bush"}: default values from Bush et al (2017) \doi{10.1038/s41467-017-00912-x} will be used
\item \code{"NA"}: all parameter will be set to \code{NA}. Usable for e.g. setting own parameter
}}
}
\value{
object of class \code{SB_strain_parameter}. The object contains a
\code{data.frame} with 8 columns and \code{n} rows. The columns are:

columns:
\itemize{
\item strain_name: the name of the strain
\item g_max_SB:
\item k_PB_SO:
\item k_SB_P:
\item h_O_SB:
\item y_SO_SB:
\item y_PB_SB:
\item m_SB:
\item i_SB:
}
}
\description{
Create new \code{SB_strain_parameter} object
}
