% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialisation_functions.R
\name{new_SB_strain_parameter}
\alias{new_SB_strain_parameter}
\title{Create SB strain parameter}
\usage{
new_SB_strain_parameter(n = 1, values = "bush")
}
\arguments{
\item{n}{number of strains}

\item{values}{values to be used or \code{"bush"}, in which case the default from Bush et al (2017) will be used}
}
\value{
data.frame with additional class \code{SB_strain_parameter}
}
\description{
Create SB strain parameter
}
