% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_strain_parameter.R
\name{new_strain_parameter}
\alias{new_strain_parameter}
\title{Creates empty parameter}
\usage{
new_strain_parameter(
  n_CB = 1,
  values_CB = "bush",
  n_PB = 1,
  values_PB = "bush",
  n_SB = 1,
  values_SB = "bush",
  other = "bush"
)
}
\arguments{
\item{n_CB}{number of strains}

\item{n_PB}{number of strains}

\item{n_SB}{number of strains}

\item{values}{values_SB to be used for SB strains parameter or \code{"bush"}, in which case the default from Bush et al (2017) will be used}

\item{values_other}{values to be used for other parameter or \code{"bush"}, in which case the default from Bush et al (2017) will be used}
}
\value{
\code{list} with additional class \code{strain_parameter}
}
\description{
The default is to create the parameter set used in Bush et al. (2017)
}
