% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_strain_parameter.R
\name{new_strain_parameter}
\alias{new_strain_parameter}
\title{Creates empty parameter}
\usage{
new_strain_parameter(n = 1)
}
\arguments{
\item{n}{number of strains or \code{"bush"}, in which case the default from Bush et al 2017 will be used}
}
\value{
\code{list} with additional class \code{strain_parameter}
}
\description{
Creates empty parameter
}
