% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dynamics_symmetric.R
\name{plot_dynamics_symmetric}
\alias{plot_dynamics_symmetric}
\title{Plot the dynamics of a model run of a symmetric parameter set}
\usage{
plot_dynamics_symmetric(simulation_result, every_n = 1, plot_a = FALSE)
}
\arguments{
\item{simulation_result}{Object returned by the run_simulation function}

\item{every_n}{Plot data of every other n sample.}

\item{plot_a}{if \code{TRUE}, diffusivities will be plotted in lower pane, default is \code{FALSE}.}
}
\value{
returns the ggplot object of the plot. If it is assigned to a
   variable, the plot needs to be plotted, otherwise it is plotted.
}
\description{
This is a convenience function to plot the dynamics of a model run particularly
for a symmetric model with strains within functional groups displayed.
}
