% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Function to run a simulation}
\usage{
run_simulation(parameter)
}
\arguments{
\item{parameter}{An object of class \code{runsim_parameter} as returned by \code{new_runsim_parameter()}.}
}
\value{
A list containing the time series of state variables and all arguments passed.
}
\description{
Function to run a simulation
}
