% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_temporal_ssfind_initial_state.R
\name{set_temporal_ssfind_initial_state}
\alias{set_temporal_ssfind_initial_state}
\title{Convenience function for setting initial states during the process of finding
stable states using the function \code{run_temporal_ssfind_experiment}
Should only be used internally.}
\usage{
set_temporal_ssfind_initial_state(
  p,
  initial_total_CB,
  initial_total_PB,
  initial_total_SB
)
}
\arguments{
\item{p}{parameters for the simulation}

\item{initial_total_CB}{Total initial abundance of CB}

\item{initial_total_PB}{Total initial abundance of PB}

\item{initial_total_SB}{Total initial abundance of SB}
}
\value{
The passed parameter object but with initial states set (overwritten by the new ones)
}
\description{
Convenience function for setting initial states during the process of finding
stable states using the function \code{run_temporal_ssfind_experiment}
Should only be used internally.
}
