% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_by_a_N.R
\name{ss_by_a_N}
\alias{ss_by_a_N}
\title{Run simulations to determine the steady states}
\usage{
ss_by_a_N(parameter, mc.cores = getOption("mc.cores", 0))
}
\arguments{
\item{parameter}{object of class \code{ss_by_a_N_parameter} as generated by \code{new_ss_by_a_N_parameter()}}

\item{mc.cores}{the number of cores to be used. If 0, the old sequential version is used.
The default is read from the option \code{mc.cores}, i.e. using \code{getOption("mc.cores", 0)}.}
}
\value{
Processed data about steady states
}
\description{
Function to get the steady states for combinations of \code{a} (oxygen diffusivity) and initial states.
This function is multithreaded, and the value of mc.cores determines the umber of parallel threads.
}
